Parse blast results
Next we will use orthomclBlastParser program to convert the tab delimited BLAST results into a format
ready for loading into the OrthoMCL schema in the relational database.
This step will also compute the percent match and percent identity of each hit. The resulting file will be
uploaded to the relational database.